+++ src/components/CartPage/CartItem.tsx
@@ -6,7 +6,7 @@ import PlusIcon from "../../assets/PlusIcon.svg?react";
 import { useRecoilState } from "recoil";
 import { Product } from "../../types";
 
-import { patchCartItemQuantity } from "../../api/cartItem";
+import { patchCartItemQuantity } from "../../api/cartItemApi";
 import { cartItemCheckedIdsAtom } from "../../recoil/atom/atom";
 import { quantitySelector } from "../../recoil/selector/selector";
 import { formatCurrency } from "../../utils/formatCurrency";
